Phase 2D: Safe Legacy Cleanup - Database Architecture Consolidation
✅ PHASE 2D SAFE LEGACY CLEANUP COMPLETE
✅ ALL 37/37 P0 TESTS PASSING - ZERO REGRESSIONS
✅ ~800 LINES OF LEGACY CODE REMOVED SAFELY

🧹 Major Phase 2D Achievements:

TYPE CONSOLIDATION:
- ✅ Extracted database types to core/database_types.py
- ✅ QueryType, WorkloadPattern, DatabaseConfig, etc. → centralized location
- ✅ Updated all imports to use core types instead of hybrid system

HYBRID MANAGER REMOVAL:
- ✅ Removed .claudedirector/lib/p0_features/shared/database_manager/ directory
- ✅ Removed 6 unused database implementation files (~800 lines)
- ✅ analytics_pipeline.py, hybrid_engine.py, semantic_search_engine.py, etc.

IMPORT MODERNIZATION:
- ✅ UnifiedDatabaseCoordinator → uses core/database_types.py
- ✅ HybridToUnifiedBridge → uses core/database_types.py
- ✅ BusinessValueCalculator → direct HybridToUnifiedBridge usage (no fallback)

SAFETY VALIDATION:
- ✅ All P0 tests passing throughout removal process
- ✅ Core database fallback safety net preserved
- ✅ UnifiedDatabaseCoordinator proven operational
- ✅ Zero functional impact on any P0 features

🏗️ Architecture Quality Improvements:
- Database types centralized in logical core location
- Reduced code duplication by ~800 lines
- Eliminated unused hybrid database implementations
- Maintained full backward compatibility during transition
- Clean separation between types and implementations

This completes Phase 2D safe cleanup while maintaining all safety nets.
Core fallback systems remain for maximum production stability.
