+++ .claudedirector/lib/p0_features/domains/strategic_metrics/business_value_calculator.py
@@ -15,19 +15,10 @@ from decimal import Decimal, ROUND_HALF_UP
 
 from ...shared.infrastructure.config import get_config
 
-# Phase 2B Migration: Use HybridToUnifiedBridge for database access
-try:
-    from ....core.hybrid_compatibility import HybridToUnifiedBridge
+# Phase 2D: Use HybridToUnifiedBridge for database access (no legacy fallback)
+from ....core.hybrid_compatibility import HybridToUnifiedBridge
 
-    HYBRID_BRIDGE_AVAILABLE = True
-    print("📊 Phase 2B: Using HybridToUnifiedBridge for BusinessValueCalculator")
-except ImportError:
-    from ...shared.database_manager.analytics_pipeline import AnalyticsPipeline
-
-    HYBRID_BRIDGE_AVAILABLE = False
-    print(
-        "📊 Phase 2B: Fallback to legacy AnalyticsPipeline for BusinessValueCalculator"
-    )
+print("📊 Phase 2D: Using HybridToUnifiedBridge for BusinessValueCalculator")
 
 logger = structlog.get_logger(__name__)
 
@@ -105,34 +96,15 @@ class BusinessValueCalculator:
     """
 
     def __init__(self, analytics_pipeline=None):
-        """Initialize BusinessValueCalculator with Phase 2B database migration support"""
-        # Phase 2B Migration: Use HybridToUnifiedBridge or fallback to AnalyticsPipeline
-        if HYBRID_BRIDGE_AVAILABLE and analytics_pipeline is None:
-            # Use HybridToUnifiedBridge for database access
-            self.database_bridge = HybridToUnifiedBridge()
-            self.analytics_pipeline = None  # Will use bridge directly
-            self.logger.info(
-                "✅ BusinessValueCalculator initialized with HybridToUnifiedBridge"
-            )
-        elif analytics_pipeline is not None:
-            # Legacy mode: use provided AnalyticsPipeline
-            self.analytics_pipeline = analytics_pipeline
-            self.database_bridge = None
-            self.logger.info(
-                "📊 BusinessValueCalculator using legacy AnalyticsPipeline"
-            )
-        else:
-            # Fallback: create bridge or raise error
-            try:
-                self.database_bridge = (
-                    HybridToUnifiedBridge() if HYBRID_BRIDGE_AVAILABLE else None
-                )
-                self.analytics_pipeline = None
-            except Exception as e:
-                raise RuntimeError(f"Failed to initialize database interface: {e}")
-
+        """Initialize BusinessValueCalculator with Phase 2D unified database architecture"""
+        # Phase 2D: Always use HybridToUnifiedBridge for database access
+        self.database_bridge = HybridToUnifiedBridge()
+        self.analytics_pipeline = None  # Unified bridge handles all analytics
+        
         self.config = get_config()
         self.logger = logger.bind(component="business_value_calculator")
+        
+        self.logger.info("✅ BusinessValueCalculator initialized with HybridToUnifiedBridge")
 
         # Business value calculation parameters
         self._value_parameters = {
