--- .claudedirector/lib/p0_features/shared/database_manager/__init__.py
@@ -1,17 +1,0 @@
-"""
-P0 Database Manager - Hybrid Architecture Framework
-
-Architecture designed for Delbert (Data Engineer) implementation.
-Provides interfaces for hybrid database management with intelligent routing.
-"""
-
-# Import only the files we successfully updated for Phase 1C
-from .analytics_pipeline import AnalyticsPipeline
-from .semantic_search_engine import SemanticSearchEngine
-from .migration_strategy import DatabaseMigrationStrategy
-
-__all__ = [
-    "AnalyticsPipeline",
-    "SemanticSearchEngine",
-    "DatabaseMigrationStrategy",
-]
