--- .claudedirector/lib/p0_features/shared/database_manager/analytics_pipeline.py
@@ -1,939 +1,0 @@
-"""
-Analytics Pipeline for Strategic Health Prediction
-
-Delbert's specialized analytics engine for health prediction AI workloads.
-Optimized for time-series analysis, trend calculation, and predictive modeling data.
-"""
-
-import time
-import json
-import sqlite3
-from typing import Dict, List, Optional, Any, Tuple, Union
-from datetime import datetime
-from dataclasses import dataclass
-from enum import Enum
-import structlog
-from concurrent.futures import ThreadPoolExecutor, as_completed
-
-from .db_base import DatabaseEngineBase
-from ....core.config import ClaudeDirectorConfig, get_config
-
-logger = structlog.get_logger(__name__)
-
-
-class AnalyticsWorkload(Enum):
-    """Types of analytics workloads for optimization"""
-
-    TIME_SERIES = "time_series"  # Temporal data analysis
-    AGGREGATION = "aggregation"  # GROUP BY, SUM, AVG operations
-    TREND_ANALYSIS = "trend_analysis"  # Pattern recognition over time
-    HEALTH_SCORING = "health_scoring"  # Initiative health calculations
-    RISK_ASSESSMENT = "risk_assessment"  # Risk factor analysis
-    CORRELATION = "correlation"  # Cross-metric relationships
-
-
-@dataclass
-class AnalyticsMetric:
-    """Individual analytics metric result"""
-
-    metric_name: str
-    value: Union[float, int, str]
-    timestamp: datetime
-    confidence: float
-    trend_direction: Optional[str] = None
-    metadata: Optional[Dict[str, Any]] = None
-
-
-@dataclass
-class AnalyticsResult:
-    """Analytics pipeline result"""
-
-    success: bool
-    workload_type: AnalyticsWorkload
-    metrics: List[AnalyticsMetric]
-    execution_time_ms: int
-    data_points_analyzed: int
-    trend_summary: Optional[Dict[str, Any]] = None
-    error: Optional[str] = None
-
-
-class AnalyticsPipeline(DatabaseEngineBase):
-    """
-    Specialized analytics pipeline for health prediction AI
-
-    Delbert's Analytics Optimizations:
-    1. Time-series data aggregation for health trends
-    2. Initiative health scoring analytics
-    3. Risk factor correlation analysis
-    4. Predictive modeling data preparation
-    5. Multi-dimensional analytics for AI training
-    """
-
-    def __init__(self, config, director_config: Optional[ClaudeDirectorConfig] = None):
-        super().__init__(config)
-        self.director_config = director_config or get_config()
-        self.logger = logger.bind(component="analytics_pipeline")
-
-        # Analytics-specific connection pool
-        self._analytics_connections: List[sqlite3.Connection] = []
-        self._pool_size = min(config.connection_pool_size, 5)  # Limit for analytics
-
-        # Analytics query patterns
-        self._analytics_patterns = {
-            "health_scoring": {
-                "initiative_progress": """
-                    SELECT initiative_key, status, risk_level,
-                           json_extract(milestone_history, '$') as milestones,
-                           json_extract(roi_tracking, '$') as roi_data,
-                           created_at, updated_at
-                    FROM strategic_initiatives
-                    WHERE status IN ('in_progress', 'at_risk', 'completed')
-                    ORDER BY updated_at DESC
-                """,
-                "stakeholder_engagement": """
-                    SELECT stakeholder_key,
-                           json_extract(interaction_history, '$') as interactions,
-                           json_extract(decision_criteria, '$') as criteria,
-                           communication_style
-                    FROM stakeholder_profiles
-                    WHERE json_extract(interaction_history, '$.interaction_count') > 0
-                """,
-                "platform_metrics": """
-                    SELECT category, metric_name, value_numeric, measurement_date,
-                           trend_direction, business_impact
-                    FROM platform_intelligence
-                    WHERE measurement_date >= date('now', '-90 days')
-                    ORDER BY measurement_date DESC, category
-                """,
-            },
-            "trend_analysis": {
-                "initiative_health_trends": """
-                    SELECT date(updated_at) as date,
-                           status,
-                           COUNT(*) as count,
-                           AVG(CASE WHEN risk_level = 'green' THEN 1.0
-                                   WHEN risk_level = 'yellow' THEN 0.5
-                                   WHEN risk_level = 'red' THEN 0.0 END) as avg_health
-                    FROM strategic_initiatives
-                    WHERE updated_at >= date('now', '-30 days')
-                    GROUP BY date(updated_at), status
-                    ORDER BY date DESC
-                """,
-                "platform_adoption_trends": """
-                    SELECT date(measurement_date) as date,
-                           category,
-                           AVG(value_numeric) as avg_value,
-                           COUNT(*) as measurement_count
-                    FROM platform_intelligence
-                    WHERE measurement_date >= date('now', '-60 days')
-                    AND value_numeric IS NOT NULL
-                    GROUP BY date(measurement_date), category
-                    ORDER BY date DESC, category
-                """,
-            },
-            "risk_assessment": {
-                "initiative_risk_factors": """
-                    SELECT initiative_key, status, risk_level,
-                           json_extract(milestone_history, '$.delays') as delays,
-                           json_extract(resource_allocation, '$.team_capacity') as capacity,
-                           (julianday('now') - julianday(created_at)) as days_active
-                    FROM strategic_initiatives
-                    WHERE status IN ('in_progress', 'at_risk')
-                """,
-                "stakeholder_risk_indicators": """
-                    SELECT s.stakeholder_key,
-                           COUNT(es.id) as recent_interactions,
-                           AVG(CASE WHEN es.persona_activated = 'crisis' THEN 1.0 ELSE 0.0 END) as crisis_ratio
-                    FROM stakeholder_profiles s
-                    LEFT JOIN executive_sessions es ON s.stakeholder_key = es.stakeholder_key
-                    WHERE es.meeting_date >= date('now', '-30 days')
-                    GROUP BY s.stakeholder_key
-                    HAVING recent_interactions > 0
-                """,
-            },
-        }
-
-        # Performance caching for analytics
-        self._analytics_cache: Dict[str, Tuple[AnalyticsResult, float]] = {}
-        self._cache_ttl = 300  # 5 minutes cache for analytics
-
-        # Health scoring weights (configurable)
-        self._health_weights = {
-            "progress_momentum": 0.25,
-            "stakeholder_satisfaction": 0.20,
-            "risk_mitigation": 0.20,
-            "timeline_adherence": 0.15,
-            "resource_efficiency": 0.10,
-            "business_impact": 0.10,
-        }
-
-    def connect(self) -> bool:
-        """Connect analytics pipeline with optimized connections"""
-        try:
-            start_time = time.time()
-
-            # Create analytics-optimized connections
-            for i in range(self._pool_size):
-                conn = self._create_analytics_connection()
-                if conn:
-                    self._analytics_connections.append(conn)
-
-            if not self._analytics_connections:
-                raise RuntimeError("Failed to create analytics connections")
-
-            # Initialize analytics schema if needed
-            self._initialize_analytics_schema()
-
-            # Create analytics-specific views
-            self._create_analytics_views()
-
-            # Pre-cache common analytics
-            self._warm_analytics_cache()
-
-            connection_time = (time.time() - start_time) * 1000
-
-            self.logger.info(
-                "Analytics pipeline connected",
-                connection_time_ms=connection_time,
-                analytics_connections=len(self._analytics_connections),
-            )
-
-            return True
-
-        except Exception as e:
-            self.logger.error("Analytics pipeline connection failed", error=str(e))
-            return False
-
-    def calculate_health_scores(
-        self, initiative_ids: Optional[List[str]] = None
-    ) -> AnalyticsResult:
-        """
-        Calculate comprehensive health scores for initiatives
-
-        Delbert's Health Scoring Algorithm:
-        1. Progress momentum analysis
-        2. Stakeholder engagement scoring
-        3. Risk factor assessment
-        4. Timeline adherence calculation
-        5. Resource efficiency analysis
-        6. Business impact measurement
-        """
-        start_time = time.time()
-
-        try:
-            # Check cache first
-            cache_key = f"health_scores_{initiative_ids or 'all'}"
-            cached_result = self._get_cached_analytics(cache_key)
-            if cached_result:
-                return cached_result
-
-            # Get analytics connection
-            conn = self._get_analytics_connection()
-
-            # Build initiative filter
-            initiative_filter = ""
-            params = []
-            if initiative_ids:
-                placeholders = ",".join("?" * len(initiative_ids))
-                initiative_filter = f"AND initiative_key IN ({placeholders})"
-                params = initiative_ids
-
-            # Execute parallel analytics queries
-            with ThreadPoolExecutor(max_workers=3) as executor:
-                # Submit analytics tasks
-                futures = {
-                    executor.submit(
-                        self._analyze_progress_momentum, conn, initiative_filter, params
-                    ): "progress",
-                    executor.submit(
-                        self._analyze_stakeholder_engagement,
-                        conn,
-                        initiative_filter,
-                        params,
-                    ): "stakeholder",
-                    executor.submit(
-                        self._analyze_risk_factors, conn, initiative_filter, params
-                    ): "risk",
-                    executor.submit(
-                        self._analyze_timeline_adherence,
-                        conn,
-                        initiative_filter,
-                        params,
-                    ): "timeline",
-                }
-
-                # Collect results
-                analytics_components = {}
-                for future in as_completed(futures):
-                    component_name = futures[future]
-                    try:
-                        analytics_components[component_name] = future.result()
-                    except Exception as e:
-                        self.logger.warning(
-                            f"Analytics component {component_name} failed", error=str(e)
-                        )
-                        analytics_components[component_name] = {}
-
-            # Calculate composite health scores
-            health_metrics = self._calculate_composite_health_scores(
-                analytics_components
-            )
-
-            execution_time = int((time.time() - start_time) * 1000)
-
-            result = AnalyticsResult(
-                success=True,
-                workload_type=AnalyticsWorkload.HEALTH_SCORING,
-                metrics=health_metrics,
-                execution_time_ms=execution_time,
-                data_points_analyzed=sum(
-                    len(comp) for comp in analytics_components.values()
-                ),
-                trend_summary=self._generate_health_trend_summary(analytics_components),
-            )
-
-            # Cache result
-            self._cache_analytics_result(cache_key, result)
-
-            self.logger.info(
-                "Health scores calculated",
-                initiatives_analyzed=len(health_metrics),
-                execution_time_ms=execution_time,
-            )
-
-            return result
-
-        except Exception as e:
-            execution_time = int((time.time() - start_time) * 1000)
-            self.logger.error("Health score calculation failed", error=str(e))
-
-            return AnalyticsResult(
-                success=False,
-                workload_type=AnalyticsWorkload.HEALTH_SCORING,
-                metrics=[],
-                execution_time_ms=execution_time,
-                data_points_analyzed=0,
-                error=str(e),
-            )
-
-    def analyze_platform_trends(self, days_back: int = 30) -> AnalyticsResult:
-        """
-        Analyze platform intelligence trends for predictive modeling
-
-        Delbert's Trend Analysis:
-        1. Platform adoption trajectory analysis
-        2. Performance metric trend detection
-        3. Cross-platform correlation analysis
-        4. Predictive trend extrapolation
-        """
-        start_time = time.time()
-
-        try:
-            cache_key = f"platform_trends_{days_back}"
-            cached_result = self._get_cached_analytics(cache_key)
-            if cached_result:
-                return cached_result
-
-            conn = self._get_analytics_connection()
-
-            # Parallel trend analysis
-            with ThreadPoolExecutor(max_workers=4) as executor:
-                futures = {
-                    executor.submit(
-                        self._analyze_adoption_trends, conn, days_back
-                    ): "adoption",
-                    executor.submit(
-                        self._analyze_performance_trends, conn, days_back
-                    ): "performance",
-                    executor.submit(
-                        self._analyze_cross_platform_correlation, conn, days_back
-                    ): "correlation",
-                    executor.submit(
-                        self._analyze_trend_velocity, conn, days_back
-                    ): "velocity",
-                }
-
-                trend_components = {}
-                for future in as_completed(futures):
-                    component_name = futures[future]
-                    try:
-                        trend_components[component_name] = future.result()
-                    except Exception as e:
-                        self.logger.warning(
-                            f"Trend component {component_name} failed", error=str(e)
-                        )
-                        trend_components[component_name] = []
-
-            # Generate trend metrics
-            trend_metrics = self._generate_trend_metrics(trend_components)
-
-            execution_time = int((time.time() - start_time) * 1000)
-
-            result = AnalyticsResult(
-                success=True,
-                workload_type=AnalyticsWorkload.TREND_ANALYSIS,
-                metrics=trend_metrics,
-                execution_time_ms=execution_time,
-                data_points_analyzed=sum(
-                    len(comp) for comp in trend_components.values()
-                ),
-                trend_summary=self._generate_platform_trend_summary(trend_components),
-            )
-
-            self._cache_analytics_result(cache_key, result)
-
-            return result
-
-        except Exception as e:
-            execution_time = int((time.time() - start_time) * 1000)
-            return AnalyticsResult(
-                success=False,
-                workload_type=AnalyticsWorkload.TREND_ANALYSIS,
-                metrics=[],
-                execution_time_ms=execution_time,
-                data_points_analyzed=0,
-                error=str(e),
-            )
-
-    def generate_ai_training_data(
-        self, workload_type: AnalyticsWorkload
-    ) -> AnalyticsResult:
-        """
-        Generate training data for AI models
-
-        Delbert's AI Data Pipeline:
-        1. Feature extraction from strategic data
-        2. Label generation for supervised learning
-        3. Data quality validation
-        4. Training/validation split preparation
-        """
-        start_time = time.time()
-
-        try:
-            conn = self._get_analytics_connection()
-            training_data = []
-
-            if workload_type == AnalyticsWorkload.HEALTH_SCORING:
-                training_data = self._extract_health_prediction_features(conn)
-            elif workload_type == AnalyticsWorkload.RISK_ASSESSMENT:
-                training_data = self._extract_risk_prediction_features(conn)
-            elif workload_type == AnalyticsWorkload.TREND_ANALYSIS:
-                training_data = self._extract_trend_prediction_features(conn)
-
-            # Convert to metrics format
-            ai_metrics = []
-            for i, data_point in enumerate(training_data):
-                metric = AnalyticsMetric(
-                    metric_name=f"training_sample_{i}",
-                    value=json.dumps(data_point),
-                    timestamp=datetime.now(),
-                    confidence=data_point.get("confidence", 1.0),
-                    metadata={
-                        "feature_count": len(data_point),
-                        "workload": workload_type.value,
-                    },
-                )
-                ai_metrics.append(metric)
-
-            execution_time = int((time.time() - start_time) * 1000)
-
-            result = AnalyticsResult(
-                success=True,
-                workload_type=workload_type,
-                metrics=ai_metrics,
-                execution_time_ms=execution_time,
-                data_points_analyzed=len(training_data),
-                trend_summary={
-                    "data_quality": self._assess_training_data_quality(training_data)
-                },
-            )
-
-            self.logger.info(
-                "AI training data generated",
-                workload_type=workload_type.value,
-                samples_generated=len(training_data),
-                execution_time_ms=execution_time,
-            )
-
-            return result
-
-        except Exception as e:
-            execution_time = int((time.time() - start_time) * 1000)
-            return AnalyticsResult(
-                success=False,
-                workload_type=workload_type,
-                metrics=[],
-                execution_time_ms=execution_time,
-                data_points_analyzed=0,
-                error=str(e),
-            )
-
-    # Private methods for analytics implementation
-
-    def _create_analytics_connection(self) -> Optional[sqlite3.Connection]:
-        """Create connection optimized for analytics workloads"""
-        try:
-            conn = sqlite3.connect(
-                str(self.config.database_path),
-                timeout=30,  # Longer timeout for analytics
-                check_same_thread=False,
-            )
-
-            conn.row_factory = sqlite3.Row
-
-            # Analytics-specific optimizations
-            conn.execute("PRAGMA journal_mode=WAL;")
-            conn.execute("PRAGMA synchronous=NORMAL;")
-            conn.execute("PRAGMA cache_size=-128000;")  # 128MB cache for analytics
-            conn.execute("PRAGMA temp_store=MEMORY;")
-            conn.execute("PRAGMA mmap_size=536870912;")  # 512MB mmap
-
-            # Analytics query optimizations
-            conn.execute("PRAGMA optimize;")
-            conn.execute("PRAGMA analysis_limit=5000;")  # Deep analysis for analytics
-
-            return conn
-
-        except Exception as e:
-            self.logger.error("Failed to create analytics connection", error=str(e))
-            return None
-
-    def _initialize_analytics_schema(self):
-        """Initialize analytics-specific schema extensions"""
-        if not self._analytics_connections:
-            return
-
-        conn = self._analytics_connections[0]
-
-        try:
-            # Create analytics summary tables if they don't exist
-            conn.execute(
-                """
-                CREATE TABLE IF NOT EXISTS analytics_cache (
-                    cache_key TEXT PRIMARY KEY,
-                    result_data TEXT NOT NULL,
-                    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                    expires_at TIMESTAMP NOT NULL
-                )
-            """
-            )
-
-            # Create analytics indexes
-            analytics_indexes = [
-                "CREATE INDEX IF NOT EXISTS idx_analytics_cache_expires ON analytics_cache(expires_at)",
-                "CREATE INDEX IF NOT EXISTS idx_initiatives_analytics ON strategic_initiatives(status, risk_level, updated_at)",
-                "CREATE INDEX IF NOT EXISTS idx_platform_analytics ON platform_intelligence(category, measurement_date, value_numeric)",
-                "CREATE INDEX IF NOT EXISTS idx_sessions_analytics ON executive_sessions(stakeholder_key, meeting_date, session_type)",
-            ]
-
-            for index_sql in analytics_indexes:
-                conn.execute(index_sql)
-
-            conn.commit()
-
-        except Exception as e:
-            self.logger.warning("Analytics schema initialization failed", error=str(e))
-
-    def _create_analytics_views(self):
-        """Create views optimized for analytics queries"""
-        if not self._analytics_connections:
-            return
-
-        conn = self._analytics_connections[0]
-
-        analytics_views = [
-            """
-            CREATE VIEW IF NOT EXISTS initiative_health_summary AS
-            SELECT
-                initiative_key,
-                status,
-                risk_level,
-                (julianday('now') - julianday(created_at)) as days_active,
-                json_extract(milestone_history, '$.completion_percentage') as progress,
-                json_extract(roi_tracking, '$.current_roi') as roi,
-                CASE
-                    WHEN risk_level = 'green' THEN 1.0
-                    WHEN risk_level = 'yellow' THEN 0.5
-                    WHEN risk_level = 'red' THEN 0.0
-                    ELSE 0.5
-                END as risk_score
-            FROM strategic_initiatives
-            WHERE status IN ('in_progress', 'at_risk', 'completed')
-            """,
-            """
-            CREATE VIEW IF NOT EXISTS platform_metrics_summary AS
-            SELECT
-                category,
-                metric_name,
-                measurement_date,
-                value_numeric,
-                trend_direction,
-                LAG(value_numeric) OVER (PARTITION BY category, metric_name ORDER BY measurement_date) as prev_value,
-                CASE
-                    WHEN trend_direction = 'improving' THEN 1.0
-                    WHEN trend_direction = 'stable' THEN 0.5
-                    WHEN trend_direction = 'declining' THEN 0.0
-                    ELSE 0.5
-                END as trend_score
-            FROM platform_intelligence
-            WHERE value_numeric IS NOT NULL
-            AND measurement_date >= date('now', '-90 days')
-            """,
-        ]
-
-        try:
-            for view_sql in analytics_views:
-                conn.execute(view_sql)
-            conn.commit()
-
-        except Exception as e:
-            self.logger.warning("Analytics views creation failed", error=str(e))
-
-    def _get_analytics_connection(self) -> sqlite3.Connection:
-        """Get analytics connection from pool"""
-        if self._analytics_connections:
-            # Rotate connections for load balancing
-            conn = self._analytics_connections.pop(0)
-            self._analytics_connections.append(conn)
-            return conn
-        else:
-            # Fallback connection
-            return self._create_analytics_connection()
-
-    def _analyze_progress_momentum(
-        self, conn: sqlite3.Connection, initiative_filter: str, params: List[str]
-    ) -> Dict[str, Any]:
-        """Analyze initiative progress momentum"""
-        query = f"""
-            SELECT initiative_key, status,
-                   json_extract(milestone_history, '$.completion_percentage') as progress,
-                   (julianday('now') - julianday(updated_at)) as days_since_update,
-                   (julianday('now') - julianday(created_at)) as total_days
-            FROM strategic_initiatives
-            WHERE status IN ('in_progress', 'at_risk', 'completed')
-            {initiative_filter}
-        """
-
-        cursor = conn.execute(query, params)
-        results = cursor.fetchall()
-
-        momentum_data = {}
-        for row in results:
-            initiative_key = row["initiative_key"]
-            progress = float(row["progress"] or 0)
-            days_since_update = float(row["days_since_update"] or 0)
-            total_days = float(row["total_days"] or 1)
-
-            # Calculate momentum score
-            velocity = progress / max(total_days, 1) if total_days > 0 else 0
-            recency_factor = max(
-                0, 1.0 - (days_since_update / 30)
-            )  # Decay over 30 days
-            momentum_score = velocity * recency_factor
-
-            momentum_data[initiative_key] = {
-                "progress": progress,
-                "velocity": velocity,
-                "recency_factor": recency_factor,
-                "momentum_score": momentum_score,
-            }
-
-        return momentum_data
-
-    def _analyze_stakeholder_engagement(
-        self, conn: sqlite3.Connection, initiative_filter: str, params: List[str]
-    ) -> Dict[str, Any]:
-        """Analyze stakeholder engagement levels"""
-        # Simplified implementation - would join with stakeholder interactions
-        query = """
-            SELECT stakeholder_key,
-                   COUNT(*) as interaction_count,
-                   AVG(CASE WHEN session_type = 'strategic_planning' THEN 1.0 ELSE 0.5 END) as engagement_weight
-            FROM executive_sessions
-            WHERE meeting_date >= date('now', '-30 days')
-            GROUP BY stakeholder_key
-        """
-
-        cursor = conn.execute(query)
-        results = cursor.fetchall()
-
-        engagement_data = {}
-        for row in results:
-            stakeholder_key = row["stakeholder_key"]
-            interaction_count = row["interaction_count"]
-            engagement_weight = float(row["engagement_weight"] or 0.5)
-
-            # Calculate engagement score
-            engagement_score = min(1.0, (interaction_count * engagement_weight) / 10)
-
-            engagement_data[stakeholder_key] = {
-                "interaction_count": interaction_count,
-                "engagement_weight": engagement_weight,
-                "engagement_score": engagement_score,
-            }
-
-        return engagement_data
-
-    def _analyze_risk_factors(
-        self, conn: sqlite3.Connection, initiative_filter: str, params: List[str]
-    ) -> Dict[str, Any]:
-        """Analyze risk factors for initiatives"""
-        query = f"""
-            SELECT initiative_key, risk_level, status,
-                   (julianday('now') - julianday(created_at)) as days_active
-            FROM strategic_initiatives
-            WHERE status IN ('in_progress', 'at_risk')
-            {initiative_filter}
-        """
-
-        cursor = conn.execute(query, params)
-        results = cursor.fetchall()
-
-        risk_data = {}
-        for row in results:
-            initiative_key = row["initiative_key"]
-            risk_level = row["risk_level"]
-            days_active = float(row["days_active"] or 0)
-
-            # Calculate risk score
-            risk_base = {"green": 0.1, "yellow": 0.5, "red": 0.9}.get(risk_level, 0.5)
-            time_factor = min(0.3, days_active / 365)  # Time-based risk increase
-            total_risk = min(1.0, risk_base + time_factor)
-
-            risk_data[initiative_key] = {
-                "risk_level": risk_level,
-                "risk_base": risk_base,
-                "time_factor": time_factor,
-                "total_risk": total_risk,
-            }
-
-        return risk_data
-
-    def _analyze_timeline_adherence(
-        self, conn: sqlite3.Connection, initiative_filter: str, params: List[str]
-    ) -> Dict[str, Any]:
-        """Analyze timeline adherence"""
-        # Simplified implementation
-        query = f"""
-            SELECT initiative_key, status,
-                   (julianday('now') - julianday(created_at)) as days_elapsed
-            FROM strategic_initiatives
-            WHERE status IN ('in_progress', 'at_risk', 'completed')
-            {initiative_filter}
-        """
-
-        cursor = conn.execute(query, params)
-        results = cursor.fetchall()
-
-        timeline_data = {}
-        for row in results:
-            initiative_key = row["initiative_key"]
-            days_elapsed = float(row["days_elapsed"] or 0)
-
-            # Simplified timeline score (would use actual target dates)
-            expected_duration = 90  # Assume 90-day average
-            adherence_score = max(
-                0, 1.0 - abs(days_elapsed - expected_duration) / expected_duration
-            )
-
-            timeline_data[initiative_key] = {
-                "days_elapsed": days_elapsed,
-                "expected_duration": expected_duration,
-                "adherence_score": adherence_score,
-            }
-
-        return timeline_data
-
-    def _calculate_composite_health_scores(
-        self, analytics_components: Dict[str, Dict[str, Any]]
-    ) -> List[AnalyticsMetric]:
-        """Calculate composite health scores from analytics components"""
-        health_metrics = []
-
-        # Get all initiative keys
-        all_initiatives = set()
-        for component_data in analytics_components.values():
-            all_initiatives.update(component_data.keys())
-
-        for initiative_key in all_initiatives:
-            # Extract component scores
-            progress_data = analytics_components.get("progress", {}).get(
-                initiative_key, {}
-            )
-            stakeholder_data = analytics_components.get("stakeholder", {}).get(
-                initiative_key, {}
-            )
-            risk_data = analytics_components.get("risk", {}).get(initiative_key, {})
-            timeline_data = analytics_components.get("timeline", {}).get(
-                initiative_key, {}
-            )
-
-            # Calculate weighted health score
-            progress_score = progress_data.get("momentum_score", 0.5)
-            stakeholder_score = stakeholder_data.get("engagement_score", 0.5)
-            risk_score = 1.0 - risk_data.get("total_risk", 0.5)  # Invert risk
-            timeline_score = timeline_data.get("adherence_score", 0.5)
-
-            # Apply weights
-            composite_score = (
-                progress_score * self._health_weights["progress_momentum"]
-                + stakeholder_score * self._health_weights["stakeholder_satisfaction"]
-                + risk_score * self._health_weights["risk_mitigation"]
-                + timeline_score * self._health_weights["timeline_adherence"]
-                + 0.7 * self._health_weights["resource_efficiency"]  # Default values
-                + 0.7 * self._health_weights["business_impact"]
-            )
-
-            # Calculate confidence based on data availability
-            data_availability = (
-                sum(
-                    [
-                        1 if progress_data else 0,
-                        1 if stakeholder_data else 0,
-                        1 if risk_data else 0,
-                        1 if timeline_data else 0,
-                    ]
-                )
-                / 4.0
-            )
-
-            confidence = max(0.5, data_availability)
-
-            # Create health metric
-            health_metric = AnalyticsMetric(
-                metric_name=f"health_score_{initiative_key}",
-                value=round(composite_score, 3),
-                timestamp=datetime.now(),
-                confidence=confidence,
-                trend_direction=self._determine_health_trend(composite_score),
-                metadata={
-                    "initiative_key": initiative_key,
-                    "component_scores": {
-                        "progress": progress_score,
-                        "stakeholder": stakeholder_score,
-                        "risk": risk_score,
-                        "timeline": timeline_score,
-                    },
-                },
-            )
-
-            health_metrics.append(health_metric)
-
-        return health_metrics
-
-    def _determine_health_trend(self, health_score: float) -> str:
-        """Determine trend direction from health score"""
-        if health_score >= 0.8:
-            return "excellent"
-        elif health_score >= 0.6:
-            return "improving"
-        elif health_score >= 0.4:
-            return "stable"
-        else:
-            return "declining"
-
-    def _generate_health_trend_summary(
-        self, analytics_components: Dict[str, Dict[str, Any]]
-    ) -> Dict[str, Any]:
-        """Generate summary of health trends"""
-        return {
-            "components_analyzed": len(analytics_components),
-            "data_quality": self.director_config.get_enum_values("priority_levels")[1],
-            "trend_confidence": 0.85,
-            "recommendation": "Continue monitoring with current metrics",
-        }
-
-    # Analytics caching methods
-
-    def _get_cached_analytics(self, cache_key: str) -> Optional[AnalyticsResult]:
-        """Get cached analytics result"""
-        if cache_key in self._analytics_cache:
-            result, timestamp = self._analytics_cache[cache_key]
-            if time.time() - timestamp < self._cache_ttl:
-                return result
-            else:
-                del self._analytics_cache[cache_key]
-        return None
-
-    def _cache_analytics_result(self, cache_key: str, result: AnalyticsResult):
-        """Cache analytics result"""
-        self._analytics_cache[cache_key] = (result, time.time())
-
-        # Limit cache size
-        if len(self._analytics_cache) > 50:
-            oldest_key = min(
-                self._analytics_cache.keys(), key=lambda k: self._analytics_cache[k][1]
-            )
-            del self._analytics_cache[oldest_key]
-
-    def _warm_analytics_cache(self):
-        """Pre-warm analytics cache with common queries"""
-        try:
-            # Pre-calculate common analytics
-            self.calculate_health_scores()
-            self.analyze_platform_trends(30)
-
-        except Exception as e:
-            self.logger.warning("Analytics cache warming failed", error=str(e))
-
-    # Placeholder methods for additional analytics
-
-    def _analyze_adoption_trends(
-        self, conn: sqlite3.Connection, days_back: int
-    ) -> List[Dict[str, Any]]:
-        """Analyze adoption trends - placeholder"""
-        return []
-
-    def _analyze_performance_trends(
-        self, conn: sqlite3.Connection, days_back: int
-    ) -> List[Dict[str, Any]]:
-        """Analyze performance trends - placeholder"""
-        return []
-
-    def _analyze_cross_platform_correlation(
-        self, conn: sqlite3.Connection, days_back: int
-    ) -> List[Dict[str, Any]]:
-        """Analyze cross-platform correlation - placeholder"""
-        return []
-
-    def _analyze_trend_velocity(
-        self, conn: sqlite3.Connection, days_back: int
-    ) -> List[Dict[str, Any]]:
-        """Analyze trend velocity - placeholder"""
-        return []
-
-    def _generate_trend_metrics(
-        self, trend_components: Dict[str, List[Dict[str, Any]]]
-    ) -> List[AnalyticsMetric]:
-        """Generate trend metrics - placeholder"""
-        return []
-
-    def _generate_platform_trend_summary(
-        self, trend_components: Dict[str, List[Dict[str, Any]]]
-    ) -> Dict[str, Any]:
-        """Generate platform trend summary - placeholder"""
-        return {"status": "placeholder"}
-
-    def _extract_health_prediction_features(
-        self, conn: sqlite3.Connection
-    ) -> List[Dict[str, Any]]:
-        """Extract features for health prediction AI - placeholder"""
-        return []
-
-    def _extract_risk_prediction_features(
-        self, conn: sqlite3.Connection
-    ) -> List[Dict[str, Any]]:
-        """Extract features for risk prediction AI - placeholder"""
-        return []
-
-    def _extract_trend_prediction_features(
-        self, conn: sqlite3.Connection
-    ) -> List[Dict[str, Any]]:
-        """Extract features for trend prediction AI - placeholder"""
-        return []
-
-    def _assess_training_data_quality(
-        self, training_data: List[Dict[str, Any]]
-    ) -> Dict[str, Any]:
-        """Assess quality of training data - placeholder"""
-        return {"quality_score": 0.85, "completeness": 0.9}
